allow config lines up to 4K length
We've run into issues with "format" in excess of the fixed buffer
size, which was 512 before this patch.  Note the updates to
"errstr" size as well in two places: I think this can sometimes
hold a config line, which is why they were sized the same before
and raised the same in this patch.

Change-Id: I4432d5305a4c7b156124db8c662e3537a98e67f5