Remove loglines cache to mitigate a possible memory leak.
Varnishkafka3 leverages a bitmap int variable returned by the Varnish 3 API
to create a cache of loglines. This behavior is not present anymore in the
Varnish 4 API and the new log grouping features filled a lot of gaps that
varnishkafka had to overcome with ad hoc code. The loglines cache is completely
removed to use a single logline, initialized at startup and re-setted after each
request proceesed (eventually freed too).

Bug: T124278
Change-Id: I40c04ca91d9d31b4d0d735298f874ed6b883120d