refactor match_assign/scratch/parser stuff
Previously, the match_assign/scratch_foo functions both called
into each other, the parsers were calling into both, and the whole
thing was rather confusing (especially with the rewinds).  This
should be cleaner and hopefully less error-prone.

There's lots of s/int/size_t/ cleanup as well, and the
less-efficient data.copy feature is removed (which defaulted on
pointlessly before, and was only to support the now-non-existent
-r option anyways).  That should make up for the doubled scratch
copies we're doing now in some !escape cases (which we're not
using in practice).

Change-Id: I057713c6b0efe7486f0163826539c5b8dfc3ec1b