Clarify the use of TAG_F_NOVARMATCH within the context of %{<strftime-format>}t.
While investigating some Timestamp inconsistencies between the old and new Varnishkafka,
I had to spend a bit of time trying to figure out the semantic of TAG_F_NOVARMATCH. Its
behavior was leading to a bullet proof result with the Varnish 3 tags since all the
timestamps had a separate and specific VSL tag, meanwhile Varnish 4 offers a more
generic SLT_Timestamp that carries the request phase (for example Start, Resp, etc..)
in its payload (together with various timings). The solution that we currently have is
working fine but it is not robust since it forces Varnishkafka to pick the first timestamp
encountered with TAG_F_NOVARMATCH, not a specific one that we set. With the current configuration,
namely grouping by Client request and discarding backend tags it works fine because the
"Start" timestamp is indeed the first one encountered, but it might need to be changed in
the future. I added a lot of comments and removed a confusing field in format_parse to
facilitate the work of whoever will refactor the code in the future. I don't see the use
case at the moment to add more code and fix this for good, but I'd really like to get
everybody's opinion to discuss what is best.

Bug: T136314
Change-Id: Icbeb5dede1cefe23a99755a68c2f0820c6c9f3e5