Add the -T VSL API timeout parameter plus the related formatter.
The Varnish workers use a buffer to write Log tags. When it gets full,
all the tags are flushed to the shmlog and anything using the VSL API
can read them. On a busy server it may happen that due to the buffer
flushed often, a Begin tag related to a request/response lands to
the shmlog but the rest is delayed (for example due to long processing).
If the delay is more than 120 seconds, the VSL API by default will return
all the available tags adding a special one to notify the issue
(VSL: timeout). Same behavior for more than 1000 pending incomplete
transactions (VSL: store overflow).

This change introduce the following improvements:
- the -T parameter to force a custom VSL timeout, rather than relying
  on the default 120 seconds.
- the -L parameter, to force an upper limit of the total VSL
  transactions to keep in memory before forcing their completion.
- a new formatter, %{VSL_API:*}x, able to log tags like VSL:timeout
  (%{VSL:*}x has been avoided since varnishncsa supports it with a
   very different semantic).

The new formatter is special since usually everything specified through fmtvar
matches a variable in the logs, namely something like "varname:". The VSL
tag does not follow this convention and puts a message directly instead.
The solution proposed is a tag flag that forces varnishkafka to use
the variable part specified in fmtvar only as string prefix, not counting the ":".
For example, a %{VSL_API:timeout}x formatter will instruct varnishkafka to search
a SLT_VSL tag with "timeout" as prefix in its payload. The special placeholder '*'
will match every error/warning message.

Change-Id: I302922a6490d287e4d866481530d2807293706a8