Improve resilience during varnish (re)starts
This patch introduces a trick already used in Varnish's VUT library,
namely forcing a reconnect to the shm log if it is detected as
abandoned or not opened. This can happen if varnishkafka starts before varnish,
or if varnish restarts/stops while varnishkafka is running. Varnishkafka is now
going to retry every 100ms endlessly until a log is aquired.

Background reading:
- https://github.com/varnishcache/varnish-cache/blob/4.1/lib/libvarnishtools/vut.c#L366
- https://github.com/varnishcache/varnish-cache/blob/master/bin/varnishlog/varnishlog.c

To keep things consistent with the current behavior, the sequence number is restarted to its
starting value when a log overrun/abandon is detected.

Bug: T138747
Change-Id: I87048786b36de325aee663a873b12874588664fb