Second submission for realloc misusage.
When using realloc all pointers are invalidated because the memory bloc may be
moved somewhere else in the address space. Thus when hitting realloc the whole
processing of the format string done before creates invalid pointers that will
cause segfaults. The realloc is called if the format string contains more than
32 elements.

Change-Id: Ia79926ab03dc4ca585fb9130ca72f7bb1c6685ed